requirements.txt
paddlepaddle==1.6.3
ple.egg==info
numpy==1.18.5
paddle==1.0.2
parl==1.3.1
